refactor(store): finish slice split — workspace.ts now 54 LOC composition
Final 4 slices land:

- view-slice (240 LOC): addView, deleteView, renameView, duplicateView,
  toggleElementInView, setLayoutDirection, resetAndRelayout, the three
  position update actions, and layoutVersion.
- tag-style-slice (83 LOC): updateElementStyle, removeElementStyle,
  renameTag, removeTagGlobal. BUILTIN_TAGS extracted to its own file
  to keep the import DAG acyclic.
- undo-slice (73 LOC): undoStack/redoStack state plus undo, redo,
  canUndo, canRedo, lastSavedUndoLength.
- lifecycle-slice (87 LOC): the workspace ref itself, loadWorkspace,
  closeWorkspace, updateWorkspaceMeta, scopeViolations, revalidateScope,
  activeWorkspaceFilename + setter.

workspace.ts is now 54 LOC: imports + slice composition + re-exports.
All 11 slices stay under 250 LOC; the previous 1216-LOC monolith is
gone. Closes the Simplicity audit MAJOR finding for src/store.

Verifications:
- npx tsc -b --force clean
- npx vitest run: 969/969
- npm run lint: 0 errors
- npx vite build: succeeds
